Develop

+++ clinic/src/main/java/com/hb/repository/impl/LabTestRepositoryImpl.java
@@ -43,7 +43,7 @@ public class LabTestRepositoryImpl extends BaseRepositoryImpl<LabTests> implemen
     }
 
     @Override
-    public LabTests getLabTestById(Integer id) {
+    public LabTests getLabTestById(Long id) {
         Session session = this.factory.getObject().getCurrentSession();
         return session.get(LabTests.class, id);
     }
@@ -59,7 +59,7 @@ public class LabTestRepositoryImpl extends BaseRepositoryImpl<LabTests> implemen
     }
 
     @Override
-    public void deleteLabTest(Integer id) {
+    public void deleteLabTest(Long id) {
         Session session = this.factory.getObject().getCurrentSession();
         LabTests test = this.getLabTestById(id);
         if (test != null) {
